Run rails new using command in dev docs
This outlines the command we should run when creating a new rails app.
https://docs.publishing.service.gov.uk/manual/setting-up-new-rails-app.html#create-a-new-rails-app

I've removed skipping action cable as we're going to be using it for
websockets.

Quite a few changes occur in this commit that are changed in
the subsequent commits, such as gems in the Gemfile, db config etc.

It might be looking at them first before digging into this commit.