Add the conversation new and create endpoints
This adds the routes, controller, and views for the new and create
conversation endpoints.

I've added a form object which handles validating the user input a
question and persisting a conversation and answer to the database.

It also adds an Error Summary Component and rendering an error message above
the textarea when the user has not input a question will be handled
in the next commit.

The Error Summary Component's implementation is tightly coupled to the
conversation form at the moment.

Down the road we may well want to refactor this into a
ErrorsHelper which handles the various mappings required for
different publishing components inputs.

See here for Whitehall's implementation:
https://github.com/alphagov/whitehall/blob/92a285e2c908a9861e1fc2d0781868263920cfae/app/helpers/errors_helper.rb#L2-L26

But for now considering this application should only have one input I think
it's more prudent to keep it simple and only add complexity if needed.