Allow conversation to be passed into form initialiser
After talking this through with Kevin, we decided that the approach
that will be used is the form being able to accept a conversation
on initialisation.

This will allow use to reuse this form when creating conversation
when the first question is asked, or append a question to a conversation.