Add validation to check all previous questions have been answered
This validation ensures that if a persisted conversation has been passed in
on initialisation, all previous questions have been answered.

It adds a scope to the question model to return all questions with
no answer present to avoid N+1 queries.

This validation is required to stop users asking multiple questions concurrently and
potentially causing issues may arise with the second response arriving
before the first etc.