Configure system specs
- Updates spec_helper to use rack_test by default
and selenium only when the `js` tag is present

We should open a PR to update govuk_test, but for
now we can use this to configure our system specs.

- updates rubocop.yml to allow instance variables in system specs

govuk-rubocop allows feature specs to use instance variables, but hasn't
been updated to allow system specs to use them.

We should update govuk-rubocop, but in the short term this allows us to
use them until we do.

- Adds config for sidekiq inline testing