Add conversations_controller#update endpoint
This adds the update endpoint which handles subsequent
questions being added to a conversation.

It re-renders the show template and dynamically sets the url of the
form based on the state of the conversation.

The request specs which check that the correct job is enqueued for a
new question has been removed based on the discussion for this PR.

This is already tested and system level and a new implementation is
being rolled out which will call the same job whichever flag is active.