Add answer component to handle rendering system generated messages
This pulls across the basic styling for system generated messages from the
prototype. I've purpsefully not worried about adding any styling used in the
JS implementation. That can be added when we start implementing the JS.

I'm keeping this in a separate commit to implementing it in the view
just so there's some chance of being able to review this PR without it
becoming one giant commit.