Add conversation show page
This adds a basic show page for conversations, which will be used to
display the questions and in a conversation and give the users the
ability to ask additional questions.

It adds some basic formatting for the users input by converting linebreaks
to br tags.

This is making no attempt to implement the JS version of the show page.