Add pending conversation endpoint
For the non-JS implementation, we need a way for a user to see that their
question has been submitted and be able to navigate back to their
conversation once the answer has been generated.

This adds the questions#pending endpoint. This redirects to the show
page if their are no pending questions on the conversation.

It provides users with a button to click to check if their answer has
been generated.

If not it re-renders the view and thanks them for their patience.

This behaviour is based on the prototype.