Use govspeak component to render answers
The message we persist in the database is in HTML. This updates the answer
component to use the govspeak component to render the answer so that the
HTML is rendered on the page for the user.

The [Govspeak component](https://components.publishing.service.gov.uk/component-guide/govspeak) ensures the GOV.UK Design System is
applied to the input.