Configure basic auth via env vars
This has been added so that we can have basic auth in a Heroku
environment.

I did write a test for this but found it was a total pain to get it to
execute reliably as we need a mechanism to reload ApplicationController
manually for test runs (as it needs to be re-run in order to pick up the
changed env vars). I decided it wasn't worth the changes (given it's
just testing a Rails tested method) but thought I'd include the tests
here in-case we wanted to use them (i.e. if basic auth went wrong).

For the tests to work I needed to change config.enable_reloading to true

spec/requests/basic_auth.rb
```
RSpec.describe "Basic auth" do
  context "when environment variables are configured to enable basic authentication" do
    around do |example|
      ClimateControl.modify(
        BASIC_AUTH_USERNAME: "username",
        BASIC_AUTH_PASSWORD: "password",
      ) do
        Rails.application.reloader.reload!
        example.run!
        Rails.application.reloader.reload!
      end
    end

    it "returns unauthorized status code when credentials aren't provided" do
      request_root
      expect(response).to have_http_status(:unauthorized)
    end

    it "returns unauthorized status code when credentials are incorrect" do
      request_root(username: "wrong", password: "also wrong")
      expect(response).to have_http_status(:unauthorized)
    end

    it "returns a successful status code when credentials are correct" do
      request_root(username: "username", password: "password")
      expect(response).to have_http_status(:success)
    end
  end

  def request_root(username: nil, password: nil)
    headers = if username || password
                credentials = Base64.encode64("#{username}:#{password}")
                { "Authorization" => "Basic #{credentials}" }
              else
                {}
              end

    get root_path
    # we expect the root path to be a redirect that needs following
    follow_redirect!(headers:)
  end
end
```