Initial Heroku configuration
This creates an app.json and a Procfile for running this application in
Heroku.

It adds the papertrail add-on for storing more logs than the measly
10-15 lines we get by default with Heroku. As well as the Postgres and
Redis logs for their respective databases.

I struggled to find good docs on how exactly this file works so mostly
got things working with a mix of trial and error. It will be best
reviewed by looking in the Heroku UI to check everything is configured
correctly.