Namespace conversation routes
Down the line we will likely want to have routes in the chat
namespace, for example a privacy policy on /chat/privacy.

Currently, the conversation show page route is chat/:id.
If we don't namespace the conversation routes, when the user hits
the hypothetical privacy page it would assume the "privacy" segement is a
UUID and blow up when it looks for the conversation with that ID.

i.e.

```
Conversation.find("privacy")
```