Use RSpec asserts over minitest view ones
This is something I missed in earlier reviews where we can use the more
idiomatic RSpec syntax.

I did feel a bit uneasy about the nested assertions where we're using a
block of an assertion, but felt it was drifting out of scope to address.
Perhaps we want to try avoid those in future?

I also had to add visibility filters for the tests that match against
the hidden sources. I'm a bit surprised it can tell that these are
hidden without rendering the CSS, I guess it's from a HTML attribute.
But anyway, they are hidden so this seems correct behaviour.

In order for requests test to work we need to add the Capybara matchers
for request tests.