Add a release process for heroku
This is the process Heroku uses when spinning up an app [1] to do any
set-up. I've configured this to run the db migrations.

I've added this in for deploying the main branch to Heroku as I can't
seem to understand how app.json can be used for that with it's
postdeploy task.

We can remove this task once we configure this app to be continuously
deployed to GOV.UK integration environment and no longer need a Heroku
environment.

[1]: https://devcenter.heroku.com/articles/release-phase#specifying-release-phase-tasks