Configure OpenSearch with Rails
This has been set-up so we can explore using OpenSearch as the means to
index GOV.UK content in a chunk form.

I've created a basic repository class that is expected to be our
applications means to interface with the index. I did struggle a bit on
naming it as we'll likely only have one index, but I did end up going
with the specific name.

Faraday Typhoeus has been added at the suggestion of the OpenSearch
README [1] for optimal performance.

It seems likely we'll have to replace our use of the opensearch-ruby gem
with the opensearch-aws-sigv4 gem and make configuration a bit more
complex for production [2].

We have a different index name for the test environment so that we can
use the same opensearch instance for development and not break the data
there when testing.

[1]: https://github.com/opensearch-project/opensearch-ruby/blob/136e1c975fc91b8cb80d7d1134e32c6dbefdb3eb/README.md#transport-features
[2]: https://github.com/opensearch-project/opensearch-ruby/blob/main/USER_GUIDE.md#amazon-opensearch-service