Use first person rather than third person user stories
The convention for BDD style stories is to write these from a first
person perspective (e.g. As a user, when I do x, then I see y). We seem
to have started by doing these in a third person style.

This can be easily corrected as we only have a few of these.