Create a message queue processor
This is a super basic message processor to just set-up the bare bones of
the handling of the message queue.

I decided to use a MessageQueue namespace instead of my earlier
suggestion of Search::Ingest (from [1]) as I felt it was unnecessary to
couple these concepts into a single namespace and felt the code would be
easier to find with this namespace. I didn't think it was necessary to
specify which message queue this code was for as there isn't likely to
be any more than one.

[1]: https://github.com/alphagov/govuk-chat/pull/58