Rake tasks to create queue and run queue consumer
This adds a pair of rake tasks which are used to create a queue and to
listen to the queue. The create queue task is expected only for use in
development environments as production will be created with terraform.
The routing key of "#" is a wildcard character that will listen to all
messages sent to the exchange. It is preferred to "*.*" as this doesn't
match routing keys with 3 segments which the Publishing API uses for
bulk re-indexing [1].

The queue task as a rake task is potentially divisive. I've followed the
instructions in govuk_message_queue_consumer [2], but I can see that
Email Alert Service removed these for a bin executable instead [3].
Therefore I imagine this approach could be subject to change as we
integrate this with the platform.

The tests for this code are extremely contrived and rely heavily on
mocks. They're basically only useful for a couple of situations: 1) APIs
of key classes change and 2) catching syntax errors in the code.

[1]: https://github.com/alphagov/publishing-api/blob/90ccece06561921fb7cc95726fe0c5d74997deaa/lib/requeue_content_by_scope.rb#L2
[1]: https://github.com/alphagov/govuk_message_queue_consumer/blob/fbf4565055d5b9122545d4c47a7187c260a94b7b/README.md#usage
[2]: https://github.com/alphagov/email-alert-service/pull/623/commits/bdf39c7330c1b494efe4ff87e4bc6ddca1d75897