Basic guard clauses on MessageProcessor
This adds guard clauses that check whether the message can be processed
or not. The basic checks here are whether the content has a base_path, as we'll
only index content with a URL, and will only index English content.

It has some basic error handling that is intended for unexpected
exceptions. I expect in the early days of this deployment there'll be a
few of those - we'll need to identify the transient ones and set-up
specific handling.

I'm using govuk_schemas as data input. Unfortunately there is a lack of
availability of notification schemas so I have to instead make use of
frontend schemas which, for our purposes, are equivalent.