Port chunking code from govuk-semantic-search
The commits that added the original code can be found in: https://github.com/alphagov/govuk-semantic-search/pull/38

These classes are being added as the logic that can take a string of
HTML content and split it into chunks based on the HTML headers.

There are few improvements that should be made to this class which I'm
going to leave until further along in this search ingestion work.