Add status and error message to answers
The status enum adds the following statuses:

- success: LLM answered the question successfully
- error_non_specific - Generic fallback for an exception situation we didn’t expect
- error_answer_service_error - Any other OpenAI error or ChatAPI error

Future statuses that will be implemented are:

- abort_forbidden_words - Question contains word from denylist
- abort_no_govuk_content - RAG returned no chunks
- error_exceeded_token_limit - exceeded the token limit

It also adds the status enum to the answer model. I've set prefix to true
as we've also done that for answer_strategy.

It also adds an error_message field which will be used to capture the error
for some states, namely error_llm_service_error & error_non_specific.