Class to represent an individual chunk of a Content Item
This class is intended to be returned as a part of an array when
converting a content item into chunks. Objects of this type are expected
to be used to populate the search index and be used to determine what
records in the search index need updating.

As the implementation is not done for populating the search index it is
somewhat presumptuous that this will meet needs and may need iteration.
The fields for what is used here is based on prototypes [1].

I implemented an inspect method as I found it was very frustrating
working with this object when debugging due to it producing a gigantic
output based on the content item.

[1]: https://github.com/kevindew/govuk-semantic-search-experiments/blob/af1a1a0e4e61ea34e8af0d63d3f071c17b7ea18b/mainstream_content/create_search_json.rb#L85-L104