Create abstract and implementation classes for content parsing
This creates the initial class that is used to convert a content item
into a series of chunks, BodyContentParser. This class deals with the
most common approach for GOV.UK content: stored in a body field in the
details hash of a content item.

I created an abstract class that has a number of helper methods which
can then be inherited. An inheriting class should implement the call
method and use helpers so it can be very succinct.

Across GOV.UK there are two different approaches for a body field. One
is a simple one where body is just a string of HTML [1], a different and
more complex one is when a body field is an array of objects that
indicate the content in different content types [2] (exclusively text/html
and text/govspeak). This class caters for both of these.

I chose to have this class raise errors for a lack of body content or
HTML content as I felt this indicated a schema we were parsing
incorrectly as the content was elsewhere. Should we learn from
exceptions that this is required I'll have to remove them.

For the tests I created a shared example which can ensure that
individual parsers conform to the API.

Approach to parsers used in this PR is inspired by Content Data API [3]

[1]: https://github.com/alphagov/publishing-api/blob/90ccece06561921fb7cc95726fe0c5d74997deaa/content_schemas/dist/formats/news_article/notification/schema.json#L394-L397
[2]: https://github.com/alphagov/publishing-api/blob/90ccece06561921fb7cc95726fe0c5d74997deaa/content_schemas/dist/formats/answer/notification/schema.json#L335-L341
[3]: https://github.com/alphagov/content-data-api/blob/37b823a188535e0385e24acb962b253918a7d8e3/app/domain/etl/edition/content/parsers/body_content.rb