Class for managing content item parsers
This class is used to determine which content item parser to call for a
particular content item and will be the interface used by the class that
synchronises the search index.

I did initially hope that we could do a loop through all the supported schemas
and confirm this class works for each one it claims to support. However
the schemas aren't precise enough in their definitions to fully match
what is needed (e.g. main text/html fields missing) so this couldn't be
relied upon.

The .supported_schemas method is intended to be used as data for a guard
clause to only accept content_items that have a schema_name value that
is supported.