Add methods to create and recreate search index
This uses the mappings that I spiked in [1]. I don't know yet if these
are all the right fields, I'm just planning to have enough to get
started with putting all the components together to have a
proof-of-concept of search ingestion. I do expect there will be a bunch
of changes.

The use of text fields is for anything we might search against and are
thus analyzed by OpenSearch, keyword fields are for exact matches only.

I may consider moving this index management to a separate class
(ChunkedContentRepository::Configuration) if the class becomes long and
unwieldy.

[1]: https://github.com/kevindew/govuk-semantic-search-experiments/blob/af1a1a0e4e61ea34e8af0d63d3f071c17b7ea18b/populate_opensearch.rb#L20-L51