Add enough code for a working Opensearch test
This just serves as a proof of concept of integrating opensearch
sufficiently that we have unit tests.

It just has one tested method on the repository class to assert we can
delete content with a base_path.

I've added the refresh_writes variable as a means that for test
environments writes can be read straight away and thus (hopefully) avoid
flaky tests due to eventual consistency. I've not got enough opensearch
experience to know if this is definitely reliable.