Add rake tasks to create and recreate index
The create task is expected to be used as part of the set-up for
govuk-docker and for anyone working on their own machine. The recreate
one provides a handy, albeit dangerous, method to quickly clear and
re-create an index.

The tests are rather contrived, they'll only really catch if the class
names or method names are changed.