Create a deterministic id for a ContentItemChunk
This id is intended to be used as the _id for index items in OpenSearch.
It is a deterministic value so that it can be used as part of upserting,
updating anything that is already using this index.

The factors to make this unique are content_id and locale to uniquely
identify the content item and then the integer chunk_index to represent
which chunk is used.

A thing to note is that this id can't be used to identify which content
was used at a particular time. For that you need to combine this id with
the digest value.