Class to synchronise a content item with search index
This is a very minimal implementation that inserts content into the
Search index with lots still to do.

I expect the actual insertion will likely move to a dedicated class as
that grows in complexity and this class will just deal with the decision
of remove content (redirects etc) or upsert content.