Replace use of frontend schemas with notification ones
I decided that using these frontend schemas weren't a good idea because
of their subtle differences to notification schema (notably around the
HTML fields) so thought I'd change these ones to be random ones.

This is somewhat the best of a bad situation type decision as these
random schemas are a very common source of flaky tests due to their
non-determinism.