Rescue standard errors that occur during answer generation
We want to ensure that we generate an answer if any error
occurs during the answer generation process. The previous commits catch
errors that are generated when making requests to OpenAI.

This commit adds error handling to the composer which will catch any standard error that
bubbles out from the OpenaiRagCompletion and GovukChatApi jobs (undefined method, etc.)

It then builds an answer with:

- a status of "error_non_specific"
- a generic message letting the user know there was an issue during generation
- sets the error_message to the class of the message and the error message
provided by the error object.

It also notifies sentry of the exception.