Add an OpenAIClient class to configure OpenAI
This class has been created to provide a central means to configure the
OpenAI::Client instance that we use.

The motivation for doing this was to:

a) Have error exceptions that were clearly from OpenAI (as opposed to
the Faraday ones)
b) Resolve an annoyance that the OpenAI gem wrote to stdout whenever an
error occurred.

point b is likely to be resolved in the next major release of the OpenAI
gem [1]. I chose not to put a guard clause on the code to remove the
middleware so that we'd get a failing test when this is gone and can
remove our code.

The idea with the wrapped errors is that we just configure ones that we
catch specifically in the application otherwise we just rely on the
general RequestError one, this is to avoid us defining lots of exception
classes that we don't use and have to keep in sync with Faraday.

I decided to name the class with the OpenAI mixed capitalisation to
match the OpenAI gem, even though it is a bit ugly to my eyes. I figure
this option is the one most likely to be applied consistently as there
is a pattern of using openai as all lower case whereas Openai looks
strange.

In testing this I struggled with trying to establish how to assert there
was not a write to stdout while catching that an error was raised. The
solution I found to this was using the RSpec negated matcher [2].

[1]: https://github.com/alexrudall/ruby-openai/pull/406
[2]: https://www.rubydoc.info/gems/rspec-expectations/3.5.0/RSpec%2FMatchers.define_negated_matcher