Add OpenAIClient::ContextLengthExceededError
This error occurs when input to a model exceeds the token length e.g:

```
[4] pry(main)> e.response[:body]
=> {"error"=>
  {"message"=>"This model's maximum context length is 4097 tokens. However, your messages resulted in 5107 tokens. Please reduce the length of the messages.",
   "type"=>"invalid_request_error",
   "param"=>"messages",
   "code"=>"context_length_exceeded"}}
```

A complication I had in putting this together was that the basic stubs
of OpenAI methods had a body of an empty string, this meant that using
the dig method would result in a TypeError. So I set up a quick
exception handler.

I borrowed the stub_openai_chat_completion_error from David's PR: https://github.com/alphagov/govuk-chat/pull/82