Consistently use OpenAI in constants
This makes the constant usages of OpenAI consistent across the app. It
consolidates that we had Openai, OpenAi and OpenAI in the app.

As an aside I'm not particularly partial to OpenAI as a constant as
mixed case ones can be quite confusing. However it does seem the best of
the available options.