Catch OpenAIClient::ContextLengthExceededError and build answer
This builds on the last commit by updating OpenAIRagCompletion to
rescue when an OpenAIClient::ContextLengthExceededError is raised.

This is caught prior to the more generic rescue that catches all
RequestError's.

When this error is caught, the answer is built using the same logic as
the existing rescue for RequestError's, but with a status of
"error_context_length_exceeded".

It also notifies sentry of the exception.