Search::TextToEmbedding class
This class is used to take a single string or an array of strings and
return back a single embedding or an array of them (if given an array).
It is expected to be used by both the code for search ingestion and
for retrieval.

It uses an approach to batch the requests into as few requests to OpenAI
as possible to provide lower latency. In my earlier tests batch requests
to OpenAI significantly reduce the latency (around a 1/10 of the time).

It's still to confirm which embedding model to use, however using the
small one likely produces better performance in early tests.

The code here is based on: [1]

[1]: https://github.com/kevindew/govuk-semantic-search-experiments/blob/af1a1a0e4e61ea34e8af0d63d3f071c17b7ea18b/mainstream_content/create_search_json.rb#L115-L149