Add a ContentSynchroniser::Result object for logging
This object is intended to be used to quickly log what has been done by
for an individual message.

Currently it just records created/deleted, however once it is expanded
to include error reasons it can then be used to move some of the ignore
conditions into the synchroniser and retain the ability to log them.

I expect that in time (likely just a few commits) we'll probably extract
the result class into a separate file.