Handle non-English content in ContentSynchroniser
An edge case is that when encountering non-English content we could be
replacing content that is in English and thus need removing from the
database. Thus rather than handling this outside the synchroniser we
should handle it.

This seem edge case applies to basically all other content checks so
this commit starts the process of moving these over to the content
synchroniser.