Add message processor guard clauses
This adds additional guard clauses to prevent indexing content that is
from schemas that aren't supported (which will also handle redirects,
gone, vanish) and content that is withdrawn.

I copied the assertion for whether content is withdrawn from [1]

Because of the fun we have with randomly generated schemas we need to
make sure any of schemas created aren't flagged as withdrawn otherwise
we get flaky tests.

[1]: https://github.com/alphagov/search-api-v2/blob/f58a885f83a22f8624f70b3c8671219d886f2714/app/models/concerns/publishing_api/action.rb#L73