Add ability to look up chunk digests
This adds a method to the ChunkedContentRepository that will return a
hash of chunk id and their corresponding digest value. The purpose of
returning this data is so we can use the digest to determine whether we
need to re-insert data into the index or not.

I did find myself with an awkward decision to make about whether to
handle pagination or not, as it's hard to imagine docs with more than
100 chunks and we could also probably push this from 100 to 1000 where
it's almost impossible.

Despite this I thought since we don't have a limit for the number of
chunks it should paginate to try handle whatever is thrown at it and not
have this part of the code introduce a surprise limit.

Should we actually decide there is a chunk limit, or if pagination
causes problems we could delete this later.