Class to index a content item
This adds and integrates ad dedicated class that determines whether the
OpenSearch index needs updating for a particular content item.

It determines what chunks already exist in the database by matching a
digest value, for any of these the index will not be updated.

It determines if there are items at the base_path with an id that does
not match the deterministic id of the chunks, these will be deleted

For other items it will add them to the index with OpenAI embeddings
applied.