Extend answer endpoint to handle "application/json" requests
For the non-js implementation when a user asks a question they are
redirected to this endpoint while the answer is being enerated.

We're adding a JS implementation which will poll this endpoint
to check if the question has been answered.

This commit updates the questions#answer endpoint to handle the requests
with the content type of "application/json". It does the following:

- if the answer has been generated respond with a 200 and the answer
- if the answer has not been generated respond with a 202 (accepted)
- if an incorrect conversation_id or question_id is provided, respond
with a 404 and the error message