BasePathVersion model for storing base_path and payload_version
This model has been created to help us achieve two tasks for message
queue consumption:

1) It allows us to know if the message we've received is new or old, as
   we cannot guarantee the order of messages received
2) It provides us with a mechanism to apply a distributed lock so that
   two queue consumer processes don't try index to the same base_path at
   the same time.

It only stores two things: a content item's base path and the payload
version number provided by the Publishing API with each message. The
base path has a unique index to ensure that we only have one record for
a particular base_path.

It is intended for this model to not store more information in future so
that it can remain only loosely coupled to the search index and is not
something that needs to be kept in sync. This will give us freedom to
easily deal with any issues with overnight syncing and ability to delete
data without painful consequences.

The payload version is necessary to be stored separate from the search
index as this also needs to keep track of content that is no longer
accessible on GOV.UK. If we were to store this in the search index it
would only be with content that is on GOV.UK and could mean that with
any out-of-order messages that we accidentally index content that is
actually removed from GOV.UK