Check messages for being out-of-date
This applies a check to the MessageProcessor to establish if messages
are received are older or newer than the data we have previously
processed. It does this by comparing the last payload_version we have
stored for that base_path with the payload_version in the message. If
the message is newer it will have a greater payload version.

For base_paths we haven't encountered before a new model will be created
with a default payload_version of 0, so this will lead to the item being
synchronised.

I decided that it would be appropriate to require a message payload
version to also be accepted if it matched the existing payload version
on the expectation that this represented data that was still current and
not out of date.