Use a distributed lock for message processing
This lock prevents two queue consumer processes operating on the same
base_path at the same time. Doing this prevents inconsistency in the
search index where two items could be trying to write different things
and leave the search index in an inconsistent state.

By using a "FOR UPDATE NOWAIT" lock this allows the first process to
claim a exclusive row level lock on the item for the duration of the
transaction. Any subsequent attempts to claim this lock will result in
an exception (this is the NOWAIT part, as the default behaviour is to
wait for the lock to be released).

It was a total pain to work out how to test this. After exploring
various ideas that involve threads that didn't lead to a proof of
concept I took the mocking way out. The evidence I have that this is the
correct exception is from testing in my console:

```
  BasePathVersion Load (15.3ms)  SELECT "base_path_versions".* FROM "base_path_versions" WHERE "base_path_versions"."id" = $1 LIMIT $2 FOR UPDATE NOWAIT  [["id", "69853003-de44-4e3a-b6cb-f9c911f03952"], ["LIMIT", 1]]
  TRANSACTION (0.3ms)  ROLLBACK
ActiveRecord::LockWaitTimeout: PG::LockNotAvailable: ERROR:  could not obtain lock on row in relation "base_path_versions"

from /root/.rbenv/versions/3.2.2/lib/ruby/gems/3.2.0/gems/activerecord-7.1.3.2/lib/active_record/connection_adapters/postgresql_adapter.rb:894:in `exec_params'
Caused by PG::LockNotAvailable: ERROR:  could not obtain lock on row in relation "base_path_versions"

from /root/.rbenv/versions/3.2.2/lib/ruby/gems/3.2.0/gems/activerecord-7.1.3.2/lib/active_record/connection_adapters/postgresql_adapter.rb:894:in `exec_params'
```