Retry MessageProcessor for service errors
This configures the MessageProcessor to retry when we experience an
error from either OpenSearch or OpenAI.

These errors are all logged to Sentry. If it turns out that some of the
errors are transient (e.g. 502 errors) and produce unactionable alerts
we should set-up a means to filter these out from Sentry.