Use factory bot to create notification content items
In the specs there are now lots of usages of GovukSchemas to generate
content items of the notification type. This code uses FactoryBot to
create these items which allow there to be a reduction in the repetitive
code used to generate these.

There's a couple of slightly quirky things I've done here that should be
explained:

Fields that are set with a default value of :preserve (as a symbol): I
wanted a means to allow these to be set to nil when running the factory
so needed a mechanism to tell FactoryBot that a particular attribute
value means that these shouldn't be set and should instead use what
comes out of the random content item generator. I settled on a symbol of
preserve with the expectation that there would not be a valid use case
to set these fields to a symbol value.

Running the validator after create: this offers an opportunity to
confirm that any schemas we modify for our tests purposes is still valid
according to the schema the Publishing API has. My logic in adding this
is that, should the Publishing API schemas change, this will help us
receive failing tests to confirm we are out of sync with the Publishing
API. I expect we need this as we have to modify the data out of the
random schema generator so heavily for tests.