Use FactoryBot to build content items in specs
This switches out the repetitive use of GovukSchemas to instead make use
of FactoryBot. This has been done for increased brevity in tests and to
allow more isolated test configuration.

For a few scenarios in BodyContentParser I had to allow schemas that
weren't valid. This doesn't strike me as a concern as this Parser class
is expected to work with a bunch of schemas.