Handle more header scenarios with HtmlHierarchicalChuker
This adds a number of extra specs for scenarios we could encounter with
headers such as: content before a header and headers out of order. It
then fixes the issues that caused those tests to fail.

The primary motivation for this was to fix a bug that we experienced on
initial tests with chunking where there was content before a header
that was not a distinct chunk.