Fix incorrect understanding of OpenAI Embedding endpoint
I had two errors in my understanding of this endpoint:

1) I thought the token limit applied to the combined input of the
   request - instead it applies per input
2) I thought that if you sent too many tokens the API would just
   truncate your input and send you embeddings for the part it
   acknowledged - instead it returns a 400 Bad Request

This commit changes this code to rectify this. We now reduce any inputs
that exceed the token length to a size within it and then send batches
to OpenAI of a fixed size.

I couldn't find any information on whether there was a limit to how many
inputs you could send OpenAI in a single request, so I ended up just
choosing an arbitrary number of 50. With an expectation that for a
chunked piece of content, it would be rare that one exceeds 50 chunks.

This uses the tiktoken gem as a means of implementing the truncation.
This gem allows the ability to convert a string into tokens and then
truncate based off the number of tokens. I was a little hesitant to use
the gem as it is flagged as in need of maintainers - but it seems to
work well.

There was some implementation options. I've maintained the approach that
we send text to OpenAI and thus we decode the string we convert to
tokens once truncated to do this. We could instead just send tokens to
OpenAI though which would remove a couple of operations - I'm just not
sure whether that carries any risks/downsides.