Add error pages in the GOV.UK Design System
This updates the application to use Rails to render errors pages in the
GOV.UK Design System.

It does the following:

- updates config.exceptions_app to use the routes
- routes the user to the errors controller
- adds an errors controller and views which handle common
exceptions (400, 403, 404, 422, 500)
- adds a request spec to test the error pages
- removes the public error pages

I've added forbidden in as we may well want to use it in future for feature
flags etc. But can remove it if people think it's unecessary for now.

I've pulled over the error messaging and styling of the main content of
the page from static so these replicate the other errors on GOV.UK.
It's likely we will want to change these down the line, but we can
use this draft PR to drive discussion.