Add abort_no_govuk_content to status enum
When we hit search and no sources are returned, we still want to respond
to the user with an answer. This answer will inform the user it in can't
find any relevant information on GOV.UK for the user.

We will build this answer with the abort_no_govuk_content status.