Use factory to generate open search chunks
This replaces the manual creation of open search chunks with the factory
created in the previoius commit.

It also updates the populate_chunked_content_index method to handle
an array of chunks being passed. Or a hash where the keys are the id
of the chunk and the value is hash that populates the data key.