Build answer sources when answer is successfully generated
When the source is generated it's url is either set to:

- the url of the Result

this is the case when only one chunk is returned for a document

- the base path of the Result

This is the case when multiple chunks are returned for a document. We only
ever want 1 source per document so we need to filter out the duplicates
when building the sources.