Use a Rails config variable for OpenAI access token
This removes the need for having to stub this env var in various places
across the test suite.

Setting the config variable as an optional env var in config/application.rb
is effective because if the env var isn't set you get a pretty unambiguous
error:

```
OpenAI::ConfigurationError:
  OpenAI access token missing! See https://github.com/alexrudall/ruby-openai#usage
```

We can then override this env var for just the test environment and not
have it affect tests further.