Run specs in a random order
This has been done as this is a  development best practice as it helps
expose issues of tests that aren't isolated and are affected by other
tests.

My motivation for adding this is that the recent issue with
https://github.com/alphagov/govuk-chat/pull/120 that concealed some
failing tests in CI. Had we had random tests we'd at least have had
flaky tests that helped alert us to this sooner.